+++ src/property-imports.ts
@@ -1,7 +1,7 @@
 import { Field } from 'contentful';
 import { ImportDeclarationStructure, OptionalKind } from 'ts-morph';
 import { linkContentTypeValidations } from './extract-validation';
-import { RenderContext } from './renderer/type';
+import { RenderContext } from './renderer';
 
 export const propertyImports = (
   field: Field,
