+++ tsconfig.json
@@ -1,11 +1,11 @@
 {
   "compilerOptions": {
     "declaration": true,
-    "importHelpers": true,
+    "importHelpers": false,
     "module": "commonjs",
     "outDir": "lib",
     "strict": true,
-    "target": "es2017",
+    "target": "esnext",
     "allowSyntheticDefaultImports": true
   },
   "include": ["src/**/*"],
